fix #34
