+++ common/src/main/java/com/adamcalculator/dynamicpack/client/ContentsScreen.java
@@ -1,7 +1,11 @@
 package com.adamcalculator.dynamicpack.client;
 
 import com.adamcalculator.dynamicpack.DynamicPackMod;
+import com.adamcalculator.dynamicpack.pack.BaseContent;
+import com.adamcalculator.dynamicpack.pack.DynamicRepoRemote;
+import com.adamcalculator.dynamicpack.pack.OverrideType;
 import com.adamcalculator.dynamicpack.pack.Pack;
+import com.adamcalculator.dynamicpack.util.Out;
 import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.ChatFormatting;
 import net.minecraft.client.Minecraft;
@@ -11,6 +15,7 @@ import net.minecraft.network.chat.CommonComponents;
 import net.minecraft.network.chat.TranslatableComponent;
 import org.jetbrains.annotations.NotNull;
 
+import java.util.LinkedHashMap;
 import java.util.function.Consumer;
 
 public class ContentsScreen extends Screen {
@@ -22,12 +27,50 @@ public class ContentsScreen extends Screen {
     private Button doneButton;
     private Button resetButton;
 
+    protected final LinkedHashMap<BaseContent, OverrideType> preChangeStates = new LinkedHashMap<>();
+
     protected ContentsScreen(Screen parent, Pack pack) {
         super(new TranslatableComponent("dynamicpack.screen.pack_contents.title"));
         this.parent = parent;
         this.pack = pack;
         this.minecraft = Minecraft.getInstance();
         this.pack.addDestroyListener(onPackReSync);
+
+        for (BaseContent knownContent : ((DynamicRepoRemote) pack.getRemote()).getKnownContents()) {
+            preChangeStates.put(knownContent, knownContent.getOverride());
+        }
+    }
+
+    public boolean isChanges() {
+        boolean t = false;
+
+        for (BaseContent knownContent : preChangeStates.keySet()) {
+            if (preChangeStates.get(knownContent) != knownContent.getOverride()) {
+                t = true;
+                break;
+            }
+        }
+        return t;
+    }
+
+    public void reset() {
+        for (BaseContent knownContent : preChangeStates.keySet()) {
+            OverrideType overrideType = preChangeStates.get(knownContent);
+            try {
+                knownContent.setOverrideType(overrideType);
+            } catch (Exception e) {
+                Out.error("Error while reset changes for content: " + knownContent, e);
+            }
+        }
+
+        contentsList.refreshAll();
+        onAfterChange();
+    }
+
+
+    public void onAfterChange() {
+        this.syncOnExit = isChanges();
+        updateDoneButton();
     }
 
     @Override
@@ -56,18 +99,13 @@ public class ContentsScreen extends Screen {
     @Override
     protected void init() {
         super.init();
-        this.contentsList = new ContentsList(this, this.minecraft, pack, (b) -> {
-            syncOnExit = b;
-            updateDoneButton(b);
-        });
-        this.addWidget(this.contentsList);
-
+        this.addWidget(this.contentsList = new ContentsList(this, this.minecraft));
         this.addRenderableWidget(doneButton = Compat.createButton(CommonComponents.GUI_DONE, this::onClose, 150, 20, this.width / 2 - 155 + 160, this.height - 29));
         this.addRenderableWidget(resetButton = Compat.createButton(new TranslatableComponent("controls.reset"), this::reset, 150, 20, this.width / 2 - 155, this.height - 29));
-        resetButton.visible = false;
+        updateDoneButton();
     }
 
-    private void updateDoneButton(boolean syncOnExit) {
+    private void updateDoneButton() {
         if (syncOnExit) {
             doneButton.setMessage(new TranslatableComponent("dynamicpack.screen.pack_contents.apply").withStyle(ChatFormatting.BOLD, ChatFormatting.GOLD));
             // TODO: 1.18.2 backport todo
@@ -79,9 +117,5 @@ public class ContentsScreen extends Screen {
             //doneButton.setTooltip(null);
         }
         resetButton.visible = syncOnExit;
-    }
-
-    private void reset() {
-        contentsList.reset();
     }
 }
